server now only outputs time , see extended coment
number of hosts, size of the input file (amount of values to be
sorted), time taken.

do note that time taken is directly related to how long it takes for
each host to attempt to connect to server. if one of the hosts take 1
day, then this time will be considered on the time output.

Time does not include loading numbers file from .txt